Seperated hotkey feature into its own class
